create CitiesContext and Reefactor App.jsx File
