+++ src/contexts/CitiesContexts.jsx
@@ -1,0 +1,35 @@
+/* eslint-disable no-unused-vars */
+import { createContext, useEffect, useState } from "react";
+const BASE_URL = "http://localhost:9000";
+
+const CitiesContext = createContext();
+
+// eslint-disable-next-line react/prop-types
+function CitiesProvider({ children }) {
+  const [cities, setCities] = useState([]);
+  const [isLoading, setIsLoading] = useState(false);
+
+  useEffect(function () {
+    async function fetchCities() {
+      try {
+        setIsLoading(true);
+        const res = await fetch(`${BASE_URL}/cities`);
+        const data = await res.json();
+        setCities(data);
+      } catch {
+        alert("There was an error Loading data");
+      } finally {
+        setIsLoading(false);
+      }
+    }
+    fetchCities();
+  }, []);
+
+  return (
+    <CitiesContext.Provider value={{ cities, isLoading }}>
+      {children}
+    </CitiesContext.Provider>
+  );
+}
+
+export { CitiesProvider };
